codeFights Skill Test
You have a box of crayons and a circle that you want to color with them.
Given a number of points on the circle and the number of crayon colors
you have, find the number of ways that you can color each point with one
of the colors such that each two adjacent points will be different
color. It is guaranteed that the answer fits in a 32-bit integer type.